+++ lib/karafka/connection/delegator.rb
@@ -2,18 +2,18 @@
 
 module Karafka
   module Connection
-    # Class that consumes messages for which we listen
-    module Processor
+    # Class that delegates processing of messages for which we listen to a proper processor
+    module Delegator
       class << self
-        # Processes messages (does something with them)
-        # It will either schedule or run a proper controller action for messages
-        # @note This should be looped to obtain a constant listening
+        # Delegates messages (does something with them)
+        # It will either schedule or run a proper processor action for messages
+        # @note This should be looped to obtain a constant delegating of new messages
         # @note We catch all the errors here, to make sure that none failures
         #   for a given consumption will affect other consumed messages
         #   If we wouldn't catch it, it would propagate up until killing the thread
         # @param group_id [String] group_id of a group from which a given message came
         # @param kafka_messages [Array<Kafka::FetchedMessage>] raw messages fetched from kafka
-        def process(group_id, kafka_messages)
+        def call(group_id, kafka_messages)
           # @note We always get messages by topic and partition so we can take topic from the
           # first one and it will be valid for all the messages
           # We map from incoming topic name, as it might be namespaced, etc.
@@ -27,7 +27,7 @@ module Karafka
           # Depending on a case (persisted or not) we might use new controller instance per each
           # batch, or use the same instance for all of them (for implementing buffering, etc)
           send(
-            topic.batch_consuming ? :process_batch : :process_each,
+            topic.batch_consuming ? :delegate_batch : :delegate_each,
             controller,
             kafka_messages
           )
@@ -35,24 +35,24 @@ module Karafka
 
         private
 
-        # Processes whole batch in one request (all at once)
+        # Delegates whole batch in one request (all at once)
         # @param controller [Karafka::BaseController] base controller descendant
         # @param kafka_messages [Array<Kafka::FetchedMessage>] raw messages from kafka
-        def process_batch(controller, kafka_messages)
+        def delegate_batch(controller, kafka_messages)
           controller.params_batch = kafka_messages
           Karafka.monitor.notice(self, kafka_messages)
           controller.call
         end
 
-        # Processes messages one by one (like with std http requests)
+        # Delegates messages one by one (like with std http requests)
         # @param controller [Karafka::BaseController] base controller descendant
         # @param kafka_messages [Array<Kafka::FetchedMessage>] raw messages from kafka
-        def process_each(controller, kafka_messages)
+        def delegate_each(controller, kafka_messages)
           kafka_messages.each do |kafka_message|
-            # @note This is a simple trick - we just process one after another, but in order
+            # @note This is a simple trick - we just delegate one after another, but in order
             # not to handle everywhere both cases (single vs batch), we just "fake" batching with
             # a single message for each
-            process_batch(controller, [kafka_message])
+            delegate_batch(controller, [kafka_message])
           end
         end
       end