+++ lib/karafka/errors.rb
@@ -39,7 +39,7 @@ module Karafka
 
     # Raised when we want to read a persisted thread messages consumer but it is unavailable
     # This should never happen and if it does, please contact us
-    MissingConsumer = Class.new(BaseError)
+    MissingClient = Class.new(BaseError)
 
     # Raised when we attemp to pause a partition but the pause timeout is equal to 0
     InvalidPauseTimeout = Class.new(BaseError)