+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - Lower ActiveSupport dependency
 - Remove configurators in favor of the after_init block configurator
 - Ruby 2.5.0 support
+- Renamed Karafka::Connection::Processor to Karafka::Connection::Delegator to match incoming naming conventions
+- Renamed Karafka::Connection::Consumer to Karafka::Connection::Client due to #274
 
 ## 1.1.2
 - #256 - Default kafka.seed_brokers configuration is created in invalid format